extract PortfolioCard component  from Portfolio
